Decimal marks in rubric are not correctly retrieved in scraping mode. Highlighter, if decimals exist in any criteria's levels they is not shown in order

+++ app/scripts/moodle/APISimulation.js
@@ -59,7 +59,7 @@ class APISimulation {
         // Get level id
         level.id = parseInt(_.last(levelElement.id.split('-')))
         // Get score
-        level.score = parseInt(levelElement.querySelector('.scorevalue').innerText)
+        level.score = parseFloat(levelElement.querySelector('.scorevalue').innerText)
         // Get descriptor
         level.definition = levelElement.querySelector('.definition').innerText
         // Get defintion format
